MNT #34 rename -- match synApps names
